requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.141.0
requests==2.32.3
pytesseract==0.3.13
Pillow==10.4.0
phonenumbers==8.13.43
email-validator==2.2.0
python-slugify==8.0.4
Jinja2==3.1.4
pytest==8.3.2
python-multipart==0.0.9
httpx==0.27.2
starlette==0.38.5
itsdangerous==2.2.0
stripe==10.9.0
